added new column Missed in daily attendance

+++ database/migrations/2021_03_15_204705_create_shifts_table.php
@@ -18,7 +18,7 @@ class CreateShiftsTable extends Migration
             $table->string('shiftName');
             $table->time('shiftStart');
             $table->time('shiftEnd');
-            $table->float('workTime');
+            $table->integer('workTime');
         });
     }
